+++ backend/garpix_company/serializers/__init__.py
@@ -1,3 +1,3 @@
 from .company import CompanySerializer, CreateCompanySerializer, UpdateCompanySerializer, ChangeOwnerCompanySerializer
-from .invite import InviteToCompanySerializer, CreateAndInviteToCompanySerializer
+from .invite import InviteToCompanySerializer, CreateAndInviteToCompanySerializer, InvitesSerializer
 from .user import GarpixCompanyUserSerializer