+++ backend/garpix_company/serializers/invite.py
@@ -1,5 +1,7 @@
+from django.conf import settings
 from django.contrib.auth import get_user_model
 from django.db import transaction
+from django.utils.module_loading import import_string
 from garpix_utils.string import get_random_string
 from rest_framework import serializers
 from rest_framework.exceptions import ValidationError
@@ -7,13 +9,20 @@ from rest_framework.exceptions import ValidationError
 from garpix_company.models.company import get_company_model
 from garpix_company.models.invite import InviteToCompany
 from django.utils.translation import ugettext_lazy as _
+from garpix_company.models.user_role import get_company_role_model
+
+
+RoleSerializer = import_string(getattr(settings, 'GARPIX_COMPANY_ROLE_SERIALIZER', 'garpix_company.serializers.role.GarpixCompanyRoleSerializer'))
 
 
 class InviteToCompanySerializer(serializers.ModelSerializer):
 
     class Meta:
         model = InviteToCompany
-        fields = ('email', 'is_admin', 'role')
+        fields = ('email', 'user', 'role')
+        extra_kwargs = {
+            'role': {'required': True},
+        }
 
     def validate_email(self, value):
         User = get_user_model()
@@ -21,25 +30,40 @@ class InviteToCompanySerializer(serializers.ModelSerializer):
         try:
             user = User.objects.get(email=value)
             if not Company.check_user_companies_limit(user):
-                raise ValidationError(_(f'У пользователя с указанным email превышен лимит количества компаний'))
+                raise ValidationError(_('У пользователя с указанным email превышен лимит количества компаний'))
         except User.DoesNotExist:
             raise ValidationError(_('Пользователь с указанным email не зарегистрирован'))
         return value
 
-    def create(self, validated_data):
+    def validate_user(self, value):
         Company = get_company_model()
-        # getting user
-        user = None
+        if not Company.check_user_companies_limit(value):
+            raise ValidationError(_('У пользователя с указанным id превышен лимит количества компаний'))
+        return value
+
+    def validate(self, data):
+        validated_data = super().validate(data)
+        user = validated_data.get('user', None)
+        email = validated_data.get('email', None)
+        if not user and not email:
+            raise ValidationError(_('Укажите email или id пользователя'))
+        if user:
+            data['email'] = user.email
+        return data
+
+    def create(self, validated_data):
+        Role = get_company_role_model()
         request = self.context.get("request")
         company_id = self.context.get("company_id")
+        if role := validated_data.get('role', None):
+            if role.role_type == Role.ROLE_TYPE.OWNER:
+                raise ValidationError({'role': [_('Нельзя пригласить пользователя на роль владельца')]})
         if request and hasattr(request, "user") and request.user.is_authenticated:
             with transaction.atomic():
                 # creating
                 obj = InviteToCompany(
-                    email=validated_data['email'],
                     company_id=company_id,
-                    is_admin=validated_data['is_admin'],
-                    role=validated_data.get('role', None)
+                    **validated_data
                 )
                 obj.save()
                 return obj
@@ -50,25 +74,31 @@ class CreateAndInviteToCompanySerializer(serializers.ModelSerializer):
 
     class Meta:
         model = InviteToCompany
-        fields = ('email', 'is_admin', 'role')
+        fields = ('email', 'role')
+        extra_kwargs = {
+            'role': {'required': True},
+        }
 
     def validate_email(self, value):
         User = get_user_model()
         if User.objects.filter(email=value).first():
-            ValidationError(_('Пользователь с указанным email уже зарегистрирован'))
+            raise ValidationError(_('Пользователь с указанным email уже зарегистрирован'))
         return value
 
     def create(self, validated_data):
         User = get_user_model()
+        Role = get_company_role_model()
         request = self.context.get("request")
         company_id = self.context.get("company_id")
+        if role := validated_data.pop('role', None):
+            if role.role_type == Role.ROLE_TYPE.OWNER:
+                raise ValidationError({'role': [_('Нельзя пригласить пользователя на роль владельца')]})
         if request and hasattr(request, "user") and request.user.is_authenticated:
             with transaction.atomic():
                 invite_data = {
                     'email': validated_data['email'],
                     'company_id': company_id,
-                    'is_admin': validated_data.pop('is_admin'),
-                    'role': validated_data.pop('role', None)
+                    'role': role
                 }
                 if 'username' not in validated_data.keys():
                     validated_data['username'] = get_random_string(25)
@@ -79,3 +109,12 @@ class CreateAndInviteToCompanySerializer(serializers.ModelSerializer):
                 obj.save()
                 return obj
         return None
+
+
+class InvitesSerializer(serializers.ModelSerializer):
+
+    role = RoleSerializer()
+
+    class Meta:
+        model = InviteToCompany
+        fields = '__all__'