+++ backend/garpix_company/serializers/role.py
@@ -1,4 +1,3 @@
-from django.contrib.auth import get_user_model
 from rest_framework import serializers
 
 from garpix_company.models.user_role import get_company_role_model
@@ -8,4 +7,4 @@ class GarpixCompanyRoleSerializer(serializers.ModelSerializer):
 
     class Meta:
         model = get_company_role_model()
-        fields = ('id', 'title')
+        fields = ('id', 'title', 'role_type')