+++ backend/garpix_company/serializers/user_company.py
@@ -6,7 +6,7 @@ from garpix_company.models.user_company import UserCompany
 
 
 UserSerializer = import_string(getattr(settings, 'GARPIX_COMPANY_USER_SERIALIZER', 'garpix_company.serializers.GarpixCompanyUserSerializer'))
-RoleSerializer = import_string(getattr(settings, 'GARPIX_COMPANY_ROLE_SERIALIZER', 'garpix_company.serializers.GarpixCompanyRoleSerializer'))
+RoleSerializer = import_string(getattr(settings, 'GARPIX_COMPANY_ROLE_SERIALIZER', 'garpix_company.serializers.role.GarpixCompanyRoleSerializer'))
 
 
 class UserCompanySerializer(serializers.ModelSerializer):
@@ -17,3 +17,10 @@ class UserCompanySerializer(serializers.ModelSerializer):
     class Meta:
         model = UserCompany
         fields = '__all__'
+
+
+class ChangeUserRoleSerializer(serializers.ModelSerializer):
+
+    class Meta:
+        model = UserCompany
+        fields = ('role',)